Fixes previous date calculation
