+++ libs/zoho-wrapper/src/zoho.ts
@@ -1,12 +1,24 @@
-import { OAuth, OAuthParams } from "./lib/auth";
+import { OAuth, OAuthParams } from './lib/auth';
+import { ZohoRequest } from './lib/requests';
+import { PerDiems } from './lib/requests/per-diems.api';
 
-export class Zoho {
-  public OAuth: OAuth;
+export interface RequestParams {
+  version?: 'v2';
+  apiHost?: string;
+  accountsHost?: string;
+  accountOwnerName: string;
+  appLinkName: string;
+}
 
-  constructor(params: OAuthParams) {
-    this.OAuth = new OAuth(params);
-    // const request = new ZohoRequest(params);
+export class ZohoApi {
+  public OAuth!: OAuth;
+  public PerDiems: PerDiems;
 
-    // this.PerDiems = new PerDiems(request);
+  // accountOwnerName, appLinkName
+  constructor(private readonly params: RequestParams) {
+    const request = new ZohoRequest(params);
+
+    this.PerDiems = new PerDiems(request);
+    // this.OAuth = new OAuth(request);
   }
 }
